stackeval: Applyable objects have a set of required components
To figure out the required order in which we'll apply the plans for all
components we need to know how data flows between them.

We'll use what's added here during the plan phase to record in the plan
which components depend on which other components, and then the apply
phase will use that coarse information to schedule the individual
component apply operations into a suitable order. None of that is
implemented here though, and will follow in future commits.